backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
chromadb==0.6.3
openai==1.59.7
anthropic==0.42.0
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.4
langchain-chroma==0.2.2
langchain-text-splitters==0.3.5
youtube-transcript-api==0.6.3
scrapetube==2.5.1
tiktoken==0.8.0
sse-starlette==2.2.1
